V.new1

+++ lib/screens/Seller/product_list_seller.dart
@@ -161,7 +161,7 @@ class _ProductListSellerState extends State<ProductListSeller> {
                           style: MyStyle().mainTitle,
                         ),
                         Text(
-                          f.format(int.parse(productModels[index].price)),
+                          f.format(double.parse(productModels[index].price)),
                           style: Theme.of(context)
                               .textTheme
                               .button!
