+++ lib/screens/User/categoryproduct.dart
@@ -177,7 +177,7 @@ class _CategoryProductState extends State<CategoryProduct> {
                           .copyWith(color: Colors.black, fontSize: 20),
                     ),
                     Text(
-                      f.format(int.parse(productModels[index].price)),
+                      f.format(double.parse(productModels[index].price)),
                       style: Theme.of(context)
                           .textTheme
                           .button!
