+++ lib/screens/User/product_user.dart
@@ -403,7 +403,7 @@ class _ProductListUserState extends State<ProductListUser> {
                             .copyWith(color: Colors.black, fontSize: 20),
                       ),
                       Text(
-                        f.format(int.parse(productRecLists[index].price)),
+                        f.format(double.parse(productRecLists[index].price)),
                         style: Theme.of(context)
                             .textTheme
                             .button!
@@ -476,7 +476,7 @@ class _ProductListUserState extends State<ProductListUser> {
                           .copyWith(color: Colors.black, fontSize: 20),
                     ),
                     Text(
-                      f.format(int.parse(productModels[index].price)),
+                      f.format(double.parse(productModels[index].price)),
                       style: Theme.of(context)
                           .textTheme
                           .button!
